+++ packages/client/ui/react-ui/src/components/hosted/v3/CrossmintHostedCheckoutV3.tsx
@@ -11,18 +11,17 @@ import { type MouseEvent, type JSX, useEffect, useState } from "react";
 
 export type CrossmintHostedCheckoutV3ReactProps = CrossmintHostedCheckoutV3Props & JSX.IntrinsicElements["button"];
 
-export function CrossmintHostedCheckout_Alpha(props: CrossmintHostedCheckoutV3ReactProps) {
+export function CrossmintHostedCheckout(props: CrossmintHostedCheckoutV3ReactProps) {
     const [didInjectCss, setDidInjectCss] = useState(false);
 
     const { crossmint } = useCrossmint();
     const apiClient = createCrossmintApiClient(crossmint);
 
     // separate custom props from jsx button props
-    const { recipient, locale, webhookPassthroughData, lineItems, payment, appearance, ...buttonProps } = props;
+    const { recipient, locale, lineItems, payment, appearance, ...buttonProps } = props;
     const customProps: CrossmintHostedCheckoutV3Props = {
         recipient,
         locale,
-        webhookPassthroughData,
         lineItems,
         payment,
         appearance,