+++ apps/auth/node-and-express-server/CHANGELOG.md
@@ -1,5 +1,17 @@
 # crossmint-auth-node
 
+## 1.1.9
+
+### Patch Changes
+
+- @crossmint/server-sdk@1.1.9
+
+## 1.1.8
+
+### Patch Changes
+
+- @crossmint/server-sdk@1.1.8
+
 ## 1.1.7
 
 ### Patch Changes