+++ packages/client/auth/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @crossmint/client-sdk-auth
 
+## 1.1.11
+
+### Patch Changes
+
+- 574f85e: Improved error throwing
+
+## 1.1.10
+
+### Patch Changes
+
+- Updated dependencies [cc64066]
+  - @crossmint/common-sdk-auth@1.0.9
+
+## 1.1.9
+
+### Patch Changes
+
+- Updated dependencies [75f80cf]
+  - @crossmint/client-sdk-base@1.4.4
+  - @crossmint/common-sdk-auth@1.0.8
+
 ## 1.1.8
 
 ### Patch Changes