+++ packages/client/window/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @crossmint/client-sdk-window
 
+## 0.2.2
+
+### Patch Changes
+
+- 75f80cf: revert: pnpm catalog
+
 ## 0.2.1
 
 ### Patch Changes