+++ packages/common/auth/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @crossmint/common-sdk-auth
 
+## 1.0.9
+
+### Patch Changes
+
+- cc64066: added twitter to auth type
+
+## 1.0.8
+
+### Patch Changes
+
+- Updated dependencies [75f80cf]
+  - @crossmint/client-sdk-base@1.4.4
+
 ## 1.0.7
 
 ### Patch Changes